+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Added - New name suffix '!' for final option.
 - Fixed - Wrong argument sometimes reported by "Unknown argument" error.
 - Added - Multivalue options (one option taking multiple arguments).
+- Fixed - opt.defaultDesc("") doesn't suppress a flag value's (default) clause.
 
 ## dimcli 7.2.0 (2024-02-22)
 - Added - More cli.to\*Cmdline() and cli.to\*CmdlineL() overloads